backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1

# LangGraph + LangChain
langgraph==0.2.60
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-openai==0.2.14
langchain-core==0.3.28

# Document Generation
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5

# Database
aiosqlite==0.20.0

# CORS
starlette==0.41.3
